@epage 2018-11-20
@emoon The power of benchmarks to identify slowdowns.

Granted, 2 of the 3 performance changes were in `render` and don't impact `render_to`.  The other is in the for_loop.  I removed an extra `clone` at the cost of less helpful error messages.